// Administrador/src/model/home.ts
export interface HomeData {
    id: number;
    name: string;
    title: string;
    description: string;
    year_experience: number;
    completed_projects: number;
    satisfied_clients: number;
    github: string;
    linkedin: string;
    instagram: string;
    x: string;
    cv: string;
    correo: string;
  }
  